fix: formatting and update dependencies

+++ src/controllers/CategoryController.ts
@@ -1,31 +1,32 @@
-import { PrismaClient } from '@prisma/client';
-import { Request, Response } from 'express'
-
-const prisma = new PrismaClient();
+import { Request, Response } from "express";
+import { prisma } from "../index";
 
 export const getAllCategory = async (req: Request, res: Response) => {
   const categories = await prisma.category.findMany();
   res.status(200).json(categories);
 };
 
-
-export const getCategoriesBySubjectId = async (req: Request, res: Response) => {
+export const getCategoryById = async (req: Request, res: Response) => {
   const { id } = req.params;
 
   try {
     const category = await prisma.category.findUnique({
       where: {
-        id: parseInt(id)
-      }
+        id: parseInt(id),
+      },
     });
 
     if (category) {
       res.status(200).json(category);
     } else {
-      res.status(404).json({ error: 'Catégorie non trouvée' });
+      res.status(404).json({ error: "Catégorie non trouvée" });
     }
   } catch (error) {
-    res.status(500).json({ message: error.message });
+    res
+      .status(500)
+      .json({
+        error: "Erreur de serveur interne. Veuillez réessayer plus tard.",
+      });
   }
 };
 
@@ -33,7 +34,7 @@ export const createCategory = async (req: Request, res: Response) => {
   const { name } = req.body;
   const newCategory = await prisma.category.create({
     data: {
-      name
+      name,
     },
   });
   res.status(200).json(newCategory);
@@ -42,12 +43,32 @@ export const createCategory = async (req: Request, res: Response) => {
 export const updateCategory = async (req: Request, res: Response) => {
   const { id } = req.params;
   const { name } = req.body;
+
+  const existingCategory = await prisma.category.findUnique({
+    where: {
+      id: parseInt(id),
+    },
+  });
+  if (!existingCategory) {
+    return res
+      .status(404)
+      .json({ error: `La catégorie à mettre à jour n'existe pas.` });
+  }
+  const existingnewCategory = await prisma.category.findFirst({
+    where: {
+      name: name,
+    },
+  });
+  if (existingnewCategory) {
+    return res.status(404).json({ error: `La catégorie à créer existe déjà.` });
+  }
+
   const updatedCategory = await prisma.category.update({
     where: {
       id: parseInt(id),
     },
     data: {
-      name
+      name,
     },
   });
   res.status(200).json(updatedCategory);
@@ -55,6 +76,19 @@ export const updateCategory = async (req: Request, res: Response) => {
 
 export const deleteCategory = async (req: Request, res: Response) => {
   const { id } = req.params;
+
+  const existingCategory = await prisma.category.findUnique({
+    where: {
+      id: parseInt(id),
+    },
+  });
+
+  if (!existingCategory) {
+    return res
+      .status(404)
+      .json({ error: `La catégorie à supprimer n'existe pas.` });
+  }
+
   await prisma.category.delete({
     where: {
       id: parseInt(id),
@@ -62,6 +96,3 @@ export const deleteCategory = async (req: Request, res: Response) => {
   });
   res.status(200).json({ message: 'Category deleted successfully' });
 };
-
-
-
