+++ src/controllers/SubjectController.ts
@@ -1,7 +1,5 @@
-import { PrismaClient } from '@prisma/client';
-import { Request, Response } from 'express'
-
-const prisma = new PrismaClient();
+import { Request, Response } from "express";
+import { prisma } from "../index";
 
 export const getAllSubject = async (req: Request, res: Response) => {
   const subjects = await prisma.subject.findMany();
@@ -14,13 +12,17 @@ export const getSubjectsByCategoryId = async (req: Request, res: Response) => {
   try {
     const subjects = await prisma.subject.findMany({
       where: {
-        categoryId: parseInt(id)
-      }
+        categoryId: parseInt(id),
+      },
     });
 
     res.status(200).json(subjects);
   } catch (error) {
-    res.status(500).json({ error: 'Erreur de serveur interne. Veuillez réessayer plus tard.' });
+    res
+      .status(500)
+      .json({
+        error: "Erreur de serveur interne. Veuillez réessayer plus tard.",
+      });
   }
 };
 
@@ -30,27 +32,55 @@ export const getSubjectsById = async (req: Request, res: Response) => {
   try {
     const subjects = await prisma.subject.findUnique({
       where: {
-        id: parseInt(id)
-      }
+        id: parseInt(id),
+      },
     });
 
     if (subjects) {
       res.status(200).json(subjects);
     } else {
-      res.status(404).json({ error: 'Matière/module non trouvée' });
+      res.status(404).json({ error: "Matière/module non trouvée" });
     }
   } catch (error) {
-    res.status(500).json({ error: 'Erreur de serveur interne. Veuillez réessayer plus tard.' });
+    res
+      .status(500)
+      .json({
+        error: "Erreur de serveur interne. Veuillez réessayer plus tard.",
+      });
   }
 };
 
 export const createSubject = async (req: Request, res: Response) => {
-  const { name, level ,category} = req.body;
+  const { name, level, categoryId } = req.body;
+
+  const existingCategory = await prisma.category.findUnique({
+    where: {
+      id: parseInt(categoryId),
+    },
+  });
+  if (!existingCategory) {
+    return res
+      .status(404)
+      .json({
+        error: `La category que vous voulez utilisé pour créer la matière/module n'existe pas.`,
+      });
+  }
+  const existingSubjectName = await prisma.subject.findFirst({
+    where: {
+      name: name,
+    },
+  });
+  if (existingSubjectName) {
+    return res
+      .status(404)
+      .json({ error: `Le nom de la matière/module à créer existe déjà` });
+  }
+
   const newSubject = await prisma.subject.create({
     data: {
       name,
       level,
-      categoryId
+      categoryId,
     },
   });
   res.status(200).json(newSubject);
@@ -60,7 +90,6 @@ export const updateSubject = async (req: Request, res: Response) => {
   const { id } = req.params;
   const { name, level, categoryId } = req.body;
 
-
   const existingSubject = await prisma.subject.findUnique({
     where: {
       id: parseInt(id),
@@ -68,7 +97,9 @@ export const updateSubject = async (req: Request, res: Response) => {
   });
 
   if (!existingSubject) {
-    return res.status(404).json({ error: `La matière/module à mettre à jour n'existe pas.` });
+    return res
+      .status(404)
+      .json({ error: `La matière/module à mettre à jour n'existe pas.` });
   }
 
   const existingCategory = await prisma.category.findUnique({
@@ -78,7 +109,11 @@ export const updateSubject = async (req: Request, res: Response) => {
   });
 
   if (!existingCategory) {
-    return res.status(404).json({ error: `La category que vous voulez utilisé pour mettre la matière/module à jour n'existe pas.` });
+    return res
+      .status(404)
+      .json({
+        error: `La category que vous voulez utilisé pour mettre la matière/module à jour n'existe pas.`,
+      });
   }
 
   const updatedSubject = await prisma.subject.update({
@@ -88,21 +123,41 @@ export const updateSubject = async (req: Request, res: Response) => {
     data: {
       name,
       level,
-      categoryId
+      categoryId,
     },
   });
   res.status(200).json(updatedSubject);
 };
 
 export const deleteSubject = async (req: Request, res: Response) => {
-  const { id } = req.params;
-  await prisma.subject.delete({
-    where: {
-      id: parseInt(id),
-    },
-  });
-  res.status(200).json({ message: 'Subject deleted successfully' });
+  try {
+    const { id } = req.params;
+
+    const existingSubject = await prisma.subject.findUnique({
+      where: {
+        id: parseInt(id),
+      },
+    });
+
+    if (!existingSubject) {
+      return res
+        .status(404)
+        .json({ error: `La matière/module à supprimer n'existe pas.` });
+    }
+
+    await prisma.subject.delete({
+      where: {
+        id: parseInt(id),
+      },
+    });
+    res
+      .status(200)
+      .json({ message: "Le module a été correctement supprimé !" });
+  } catch (error) {
+    res
+      .status(500)
+      .json({
+        error: "Une erreur est survenue lors de la suppression du module.",
+      });
+  }
 };
-
-
-
